Rearrange folders and paths - introduce new alias in webpack for easy access to new folders
