+++ src/core/reducers/globalActivities.reducer.test.js
@@ -1,8 +1,8 @@
 jest.mock('../services/LocalStorage')
 
-import storage from '../services/LocalStorage'
-import globalActivitiesReducer from 'src/reducers/globalActivities.reducer'
-import { setActivities } from 'src/actions/projectIssues.action'
+import storage from 'src/services/LocalStorage'
+import globalActivitiesReducer from 'core/reducers/globalActivities.reducer'
+import { setActivities } from 'core/actions'
 
 const MOCK_ACTIVITIES_WITH_DEFAULT = [
   { id: 8, name: 'Design' },
