+++ src/web/assets/custom.css
@@ -82,7 +82,7 @@ h5 {
 .rfh-time-log-button {
   width: 100%;
   border: 0;
-  transition: all .2s ease-in-out;
+  transition: all 0.2s ease-in-out;
   color: white;
   /* Permalink - use to edit and share this gradient: http://colorzilla.com/gradient-editor/#d4145a+0,f26851+100 */
   background: #d4145a; /* Old browsers */
