+++ src/web/index.js
@@ -5,14 +5,14 @@ import { render } from 'react-dom'
 import Bootstrap from 'bootstrap/dist/css/bootstrap.css'
 import { createLogger } from 'redux-logger'
 import thunk from 'redux-thunk'
-import userLogTimeEntries from 'src/reducers/userLogTimeEntries.reducer'
-import authentication from 'src/reducers/authentication.reducer'
-import projectIssues from 'src/reducers/projectIssues.reducer'
-import projects from 'src/reducers/projects.reducer'
-import globalActivities from 'src/reducers/globalActivities.reducer'
-import user from 'src/reducers/user.reducer'
-import Root from './Root.jsx'
-import styles from './assets/custom.css'
+import userLogTimeEntries from 'core/reducers/userLogTimeEntries.reducer'
+import authentication from 'core/reducers/authentication.reducer'
+import projectIssues from 'core/reducers/projectIssues.reducer'
+import projects from 'core/reducers/projects.reducer'
+import globalActivities from 'core/reducers/globalActivities.reducer'
+import user from 'core/reducers/user.reducer'
+import Root from 'web/Root.jsx'
+import styles from 'web/assets/custom.css'
 
 const reducers = combineReducers({
   projectIssues,
