+++ webpack.desktop.production.config.js
@@ -5,7 +5,7 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 module.exports = {
   mode: 'production',
   target: 'electron-renderer',
-  entry: './src/index.js',
+  entry: './src/web/index.js',
   output: {
     path: resolve(__dirname, 'public'),
     filename: 'bundle.js'
